// web/generated/schema.ts
export interface Menu extends SanityDocument {
  _type: "menu"

  /**
   * Title — `string`
   *
   *
   */
  title?: string

  /**
   * Manu Items — `array`
   *
   *
   */
  items?: Array<SanityKeyedReference<Page>>
}

/**
 * Page
 *
 *
 */
export interface Page extends SanityDocument {
  _type: "page"

  /**
   * Title — `localeString`
   *
   *
   */
  title?: LocaleString

  /**
   * Menu Title — `localeString`
   *
   * The title shown in the main navigation
   */
  menuTitle?: LocaleString

  /**
   * Subtitle — `localeString`
   *
   * Not available for 'Page' type
   */
  subtitle?: LocaleString

  /**
   * Template — `array`
   *
   *
   */
  template?: Array<SanityKeyed<string>>

  /**
   * Slug — `localeSlug`
   *
   *
   */
  slug?: LocaleSlug

  /**
   * Body — `localeRichText`
   *
   * Only available for 'Page' type
   */
  body?: LocaleRichText

  /**
   * Main image — `captionImage`
   *
   * Only available for 'Page' type
   */
  mainImage?: CaptionImage

  /**
   * SEO title — `localeString`
   *
   * Displayed on Facebook and Twitter shares (max 60 characters).
   */
  seoTitle?: LocaleString

  /**
   * SEO description — `localeString`
   *
   * Displayed on Facebook and Twitter shares (max 65 characters).
   */
  seoDescription?: LocaleString

  /**
   * SEO Image — `image`
   *
   * Ideal size 1200 x 630px (if not added main image will be used).
   */
  seoImage?: {
    _type: "image"
    asset: SanityReference<SanityImageAsset>
    crop?: SanityImageCrop
    hotspot?: SanityImageHotspot
  }
}
// ...
export type BlockContent = Array<
  | SanityKeyed<SanityBlock>
  | SanityKeyed<{
      _type: "image"
      asset: SanityReference<SanityImageAsset>
      crop?: SanityImageCrop
      hotspot?: SanityImageHotspot
    }>
>

export type CaptionImage = {
  _type: "image"
  asset: SanityReference<SanityImageAsset>
  crop?: SanityImageCrop
  hotspot?: SanityImageHotspot

  /**
   * Caption — `string`
   *
   *
   */
  caption?: string

  /**
   * Attribution — `string`
   *
   *
   */
  Caption?: string
}

export type LocaleRichText = {
  _type: "localeRichText"
  /**
   * English — `blockContent`
   *
   *
   */
  en?: BlockContent

  /**
   * Welsh — `blockContent`
   *
   *
   */
  cy?: BlockContent
}

export type LocaleSlug = {
  _type: "localeSlug"
  /**
   * English — `slug`
   *
   *
   */
  en?: { _type: "en"; current: string }

  /**
   * Welsh — `slug`
   *
   *
   */
  cy?: { _type: "cy"; current: string }
}

export type LocaleString = {
  _type: "localeString"
  /**
   * English — `string`
   *
   *
   */
  en?: string

  /**
   * Welsh — `string`
   *
   *
   */
  cy?: string
}